Move some converters to core, attribute the types that use them there
