+++ src/Core/src/Converters/CornerRadiusTypeConverter.cs
@@ -2,9 +2,8 @@ using System;
 using System.ComponentModel;
 using System.Globalization;
 
-namespace Microsoft.Maui.Controls
+namespace Microsoft.Maui.Converters
 {
-	[Xaml.TypeConversion(typeof(CornerRadius))]
 	public class CornerRadiusTypeConverter : TypeConverter
 	{
 		public override bool CanConvertFrom(ITypeDescriptorContext context, Type sourceType)
